+++ marketPlace/src/main/java/com/mktpc/marketPlace/model/dtos/ProductDTO.java
@@ -2,5 +2,5 @@ package com.mktpc.marketPlace.model.dtos;
 
 import com.mktpc.marketPlace.model.CategoryProduct;
 
-public record ProductDTO (String name, String productDetails, CategoryProduct categoryProduct, int quant, double priceProduct) {
+public record ProductDTO (String name, String productDetails, CategoryProduct categoryProduct, Long quant, double priceProduct) {
 }
